# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Meu Auto API
  version: "1.0.0"
  description: |
    API do Meu Auto — acompanhamento da vida útil de um veículo para o dono pessoa física
    no Brasil.

    **Convenções**

    - Datas civis (dia de uma manutenção, vencimento de IPVA) são `YYYY-MM-DD`, sem hora e
      sem fuso. Timestamps de auditoria (`created_at`, `updated_at`) são RFC 3339.
    - Dinheiro é sempre **inteiro em centavos** (`*_cents`). Nunca float, nunca string.
    - Distância é sempre inteiro em quilômetros.
    - Mensagens de erro são pt-BR e podem ser exibidas ao usuário. O `code` é estável e em
      inglês — é nele que o cliente decide comportamento, nunca na mensagem.
    - `POST` de recurso aceita `id` gerado pelo cliente (UUIDv7). Reenviar a mesma
      requisição é idempotente: retorna `200` com o recurso já criado em vez de `201`.
    - Listas paginadas usam cursor opaco (`next_cursor`), nunca offset. `next_cursor: null`
      é a última página. O cursor é emitido pelo servidor — não construa um.
  contact:
    name: Meu Auto
  license:
    name: Proprietary
    url: https://meuauto.com.br
servers:
  - url: http://localhost:8080
    description: Desenvolvimento local
  - url: https://{host}
    description: Ambiente hospedado
    variables:
      host:
        default: api.meuauto.com.br
...
security:
  - bearerAuth: []
paths:
...
  /v1/maintenance-items:
    get:
      operationId: listMaintenanceItems
      tags: [manutencao]
      summary: Catálogo de manutenção
      description: Itens globais mais os itens personalizados do próprio usuário.
      parameters:
        - name: vehicle_type
          in: query
          schema: { $ref: "#/components/schemas/VehicleType" }
        - name: kind
          in: query
          schema: { $ref: "#/components/schemas/MaintenanceItemKind" }
      responses:
        "200":
          description: O catálogo visível ao usuário
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/MaintenanceItem" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    post:
      operationId: createMaintenanceItem
      tags: [manutencao]
      summary: Criar item personalizado
      description: Visível apenas para quem criou. O `slug` é derivado do nome.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMaintenanceItemRequest" }
      responses:
        "201":
          description: Item criado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaintenanceItem" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /v1/vehicles/{vehicleId}/maintenance-plans:
    parameters:
      - $ref: "#/components/parameters/VehicleId"
    get:
      operationId: listMaintenancePlans
      tags: [manutencao]
      summary: Planos de manutenção, com o vencimento já calculado
      description: |
        Ordenado por urgência: vencido, vence em breve, sem baseline, em dia, sem
        periodicidade.

        `due_at_km`/`remaining_km` são `null` quando o plano não tem intervalo por
        distância; `due_on`/`remaining_days` são `null` quando não tem intervalo por tempo.
        `null` significa **essa dimensão não se aplica**, nunca zero.
      responses:
        "200":
          description: Os planos do veículo
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/MaintenancePlan" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      operationId: createMaintenancePlan
      tags: [manutencao]
      summary: Criar plano
      description: |
        Se nenhum intervalo for informado, os padrões do item no catálogo são usados —
        criar um plano pode ser só nomear o item.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMaintenancePlanRequest" }
      responses:
        "201":
          description: Plano criado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaintenancePlanSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /v1/maintenance-plans/{planId}:
    parameters:
      - name: planId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updateMaintenancePlan
      tags: [manutencao]
      summary: Atualizar plano
      description: |
        Qualquer edição marca o plano como `origin: "user"`, para que uma futura atualização
        dos padrões sugeridos não sobrescreva a customização.

        Para transformar o plano em um que só agrupa histórico e nunca vence, envie
        `clear_intervals: true` — sem intervalos junto.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateMaintenancePlanRequest" }
      responses:
        "200":
          description: Plano atualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaintenancePlanSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    delete:
      operationId: deleteMaintenancePlan
      tags: [manutencao]
      summary: Desativar plano
      description: Desativa em vez de apagar — o plano pode já ter histórico associado.
      responses:
        "204": { description: Plano desativado }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/vehicles/{vehicleId}/maintenance-records:
    parameters:
      - $ref: "#/components/parameters/VehicleId"
    get:
      operationId: listMaintenanceRecords
      tags: [manutencao]
      summary: Manutenções realizadas
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Uma página do histórico, mais recente primeiro
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaintenanceRecordPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    post:
      operationId: createMaintenanceRecord
      tags: [manutencao]
      summary: Registrar manutenção
      description: |
        Um registro é o evento (data, quilometragem, oficina, valor total) com **uma ou
        mais linhas de item**. É esse formato que faz uma "revisão dos 100.000 km"
        funcionar: um evento, uma nota fiscal, e cada item zerando o próprio relógio.

        Pelo menos um item é obrigatório — um registro sem item não zera relógio nenhum.
        Use o item `personalizada` do catálogo para o que não tem nome.

        A quilometragem informada gera uma leitura de odômetro e passa pela mesma validação
        de consistência do endpoint de odômetro.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMaintenanceRecordRequest" }
      responses:
        "200":
          description: O `id` enviado já existia neste veículo — retentativa
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaintenanceRecord" }
        "201":
          description: Manutenção registrada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaintenanceRecord" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422":
          description: Dados inválidos, ou quilometragem inconsistente com o histórico
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/maintenance-records/{recordId}:
    parameters:
      - name: recordId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getMaintenanceRecord
      tags: [manutencao]
      summary: Ver manutenção
      responses:
        "200":
          description: A manutenção, com suas linhas
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaintenanceRecord" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateMaintenanceRecord
      tags: [manutencao]
      summary: Atualizar manutenção
      description: |
        Altera data, quilometragem, oficina, valor e observação. **Não altera as linhas de
        item** — trocar item a item exige decidir o que acontece com o relógio dos removidos.

        Mudar data ou quilometragem move a leitura de odômetro gerada por este registro, e
        passa pela mesma validação de consistência.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateMaintenanceRecordRequest" }
      responses:
        "200":
          description: Manutenção atualizada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaintenanceRecord" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    delete:
      operationId: deleteMaintenanceRecord
      tags: [manutencao]
      summary: Retratar manutenção
      description: |
        Exclusão lógica. A leitura de odômetro gerada por este registro é removida junto e o
        relógio dos itens envolvidos volta ao registro anterior — um serviço retratado não
        pode continuar alimentando o histórico com um número que ninguém sustenta.
      responses:
        "204": { description: Manutenção retratada }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------------------------------------------------------------- prazos
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token de acesso obtido em `/v1/auth/login`, `/register` ou `/refresh`. Vale 15
        minutos; renove com o refresh token antes de expirar.
  parameters:
    VehicleId:
      name: vehicleId
      in: path
      required: true
      schema: { type: string, format: uuid }
    Limit:
      name: limit
      in: query
      description: Tamanho da página. Valores fora da faixa são ajustados, não rejeitados.
      schema: { type: integer, format: int32, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      name: cursor
      in: query
      description: |
        Cursor opaco devolvido em `next_cursor`. Repasse sem interpretar — o formato é
        interno e pode mudar.
      schema: { type: string }
  responses:
    Unauthorized:
      description: Autenticação necessária, ou sessão inválida/expirada
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: |
        Recurso não encontrado.

        Também é a resposta para um recurso que existe mas não pertence ao usuário: dizer
        "você não pode ver isto" confirmaria que ele existe.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Conflito com um recurso existente
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationFailed:
      description: |
        Dados inválidos. `details.fields` traz **todos** os campos rejeitados de uma vez,
        com a chave igual à do JSON enviado.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: validation_failed
              message: Não foi possível criar a conta.
              details:
                fields:
                  email: Informe um e-mail válido.
                  password: A senha deve ter pelo menos 8 caracteres.
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { $ref: "#/components/schemas/ErrorCode" }
            message:
              type: string
              description: Mensagem em pt-BR, exibível ao usuário.
            details:
              type: object
              additionalProperties: true
              description: |
                Dados estruturados específicos do erro. Em `validation_failed` traz
                `fields`; em `odometer_rollback`, o registro vizinho conflitante; em
                `internal`, o `request_id` para suporte.
    ErrorCode:
      type: string
      description: |
        Identificador estável do erro. **É nele que o cliente decide comportamento**, nunca
        na mensagem.

        Valores podem ser acrescentados no futuro — trate um código desconhecido como falha
        genérica, nunca como erro de parsing.
      enum:
        - validation_failed
        - unauthorized
        - forbidden
        - not_found
        - method_not_allowed
        - conflict
        - odometer_rollback
        - rate_limited
        - internal

    # ------------------------------------------------------------ enums de domínio
    VehicleType:
      type: string
      description: MVP-1 aceita apenas `car`.
      enum: [car]
...
    MaintenanceItemKind:
      type: string
      description: |
        `maintenance` é histórico de serviço; `care` é hábito recorrente (calibrar pneus,
        lavar o carro). Ambos passam pelo mesmo motor de vencimento.
      enum: [maintenance, care]
    MaintenanceRecordKind:
      type: string
      description: |
        `performed` é um serviço que aconteceu e pode ser comprovado. `declared` é o dono
        afirmando de memória — o carro usado comprado com "última troca aos 95.000 km" e sem
        recibo.
      enum: [performed, declared]
    MaintenancePlanOrigin:
      type: string
      description: |
        `suggested` veio do catálogo no cadastro do veículo; `user` foi criado ou editado
        pelo dono. Qualquer edição promove para `user`.
      enum: [suggested, user]
    MaintenanceStatus:
      type: string
      description: |
        Estado de um plano. Calculado a cada leitura, nunca armazenado.

        - `vencido` — algum limite foi ultrapassado.
        - `vence_em_breve` — algum limite está dentro da antecedência configurada.
        - `em_dia` — nada perto.
        - `sem_baseline` — existe regra, mas nada foi registrado ainda. Não é alerta: é
          pedido de configuração.
        - `sem_periodicidade` — o plano agrupa histórico e nunca vence. Estado válido, não
          dado faltando.
      enum: [vencido, vence_em_breve, em_dia, sem_baseline, sem_periodicidade]
...
    CreateMaintenanceItemRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 120 }
        kind: { $ref: "#/components/schemas/MaintenanceItemKind" }
        default_interval_km: { type: integer, format: int32, nullable: true, minimum: 1 }
        default_interval_months: { type: integer, format: int32, nullable: true, minimum: 1 }
        default_interval_days: { type: integer, format: int32, nullable: true, minimum: 1 }
    MaintenanceItem:
      type: object
      required: [id, slug, name, kind, vehicle_type, is_custom]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        kind: { $ref: "#/components/schemas/MaintenanceItemKind" }
        vehicle_type: { type: string, enum: [car, motorcycle, all] }
        is_custom:
          type: boolean
          description: '`true` quando é um item criado por este usuário, não do catálogo global.'
        default_interval_km: { type: integer, format: int32, nullable: true }
        default_interval_months: { type: integer, format: int32, nullable: true }
        default_interval_days: { type: integer, format: int32, nullable: true }
    CreateMaintenancePlanRequest:
      type: object
      required: [maintenance_item_id]
      properties:
        id: { type: string, format: uuid, nullable: true }
        maintenance_item_id: { type: string, format: uuid }
        interval_km: { type: integer, format: int32, nullable: true, minimum: 1 }
        interval_months: { type: integer, format: int32, nullable: true, minimum: 1 }
        interval_days: { type: integer, format: int32, nullable: true, minimum: 1 }
        alert_km: { type: integer, format: int32, nullable: true, minimum: 0 }
        alert_days: { type: integer, format: int32, nullable: true, minimum: 0 }
    UpdateMaintenancePlanRequest:
      type: object
      properties:
        interval_km: { type: integer, format: int32, nullable: true, minimum: 1 }
        interval_months: { type: integer, format: int32, nullable: true, minimum: 1 }
        interval_days: { type: integer, format: int32, nullable: true, minimum: 1 }
        alert_km: { type: integer, format: int32, nullable: true, minimum: 0 }
        alert_days: { type: integer, format: int32, nullable: true, minimum: 0 }
        clear_intervals:
          type: boolean
          description: |
            Limpa os três intervalos, transformando o plano em um que só agrupa histórico e
            nunca vence. Precisa de flag própria porque `null` já significa "não altere".
            Não pode vir junto de um intervalo.
    MaintenancePlanSummary:
      type: object
      description: O plano sem o estado calculado. Devolvido em criação e atualização.
      required: [id, maintenance_item_id, alert_km, alert_days, origin]
      properties:
        id: { type: string, format: uuid }
        maintenance_item_id: { type: string, format: uuid }
        interval_km: { type: integer, format: int32, nullable: true }
        interval_months: { type: integer, format: int32, nullable: true }
        interval_days: { type: integer, format: int32, nullable: true }
        alert_km: { type: integer, format: int32 }
        alert_days: { type: integer, format: int32 }
        origin: { $ref: "#/components/schemas/MaintenancePlanOrigin" }
    MaintenancePlan:
      type: object
      description: O plano com o vencimento já calculado.
      required:
        [id, maintenance_item_id, item_slug, item_name, item_kind, alert_km, alert_days, origin, status]
      properties:
        id: { type: string, format: uuid }
        maintenance_item_id: { type: string, format: uuid }
        item_slug: { type: string }
        item_name: { type: string }
        item_kind: { $ref: "#/components/schemas/MaintenanceItemKind" }
        interval_km: { type: integer, format: int32, nullable: true }
        interval_months: { type: integer, format: int32, nullable: true }
        interval_days: { type: integer, format: int32, nullable: true }
        alert_km: { type: integer, format: int32 }
        alert_days: { type: integer, format: int32 }
        origin: { $ref: "#/components/schemas/MaintenancePlanOrigin" }
        status: { $ref: "#/components/schemas/MaintenanceStatus" }
        due_at_km:
          type: integer
          format: int32
          nullable: true
          description: '`null` quando o plano não tem intervalo por distância.'
        due_on:
          type: string
          format: date
          nullable: true
          description: '`null` quando o plano não tem intervalo por tempo.'
        remaining_km: { type: integer, format: int32, nullable: true }
        remaining_days: { type: integer, format: int32, nullable: true }
        last_occurred_on:
          type: string
          format: date
          nullable: true
          description: Data do último registro deste item. `null` quando `status` é `sem_baseline`.
        last_mileage_km: { type: integer, format: int32, nullable: true }
    MaintenanceRecordItemRequest:
      type: object
      required: [maintenance_item_id]
      properties:
        maintenance_item_id: { type: string, format: uuid }
        description: { type: string, nullable: true }
        part_brand: { type: string, nullable: true }
        cost_cents: { type: integer, format: int64, nullable: true, minimum: 0 }
        warranty_months: { type: integer, format: int32, nullable: true, minimum: 1 }
        warranty_km: { type: integer, format: int32, nullable: true, minimum: 1 }
    CreateMaintenanceRecordRequest:
      type: object
      required: [mileage_km, items]
      properties:
        id: { type: string, format: uuid, nullable: true }
        occurred_on:
          type: string
          format: date
          nullable: true
          description: Padrão é hoje. Não pode estar no futuro.
        mileage_km: { type: integer, format: int32, minimum: 0 }
        kind: { $ref: "#/components/schemas/MaintenanceRecordKind" }
        workshop_name: { type: string, nullable: true, maxLength: 120 }
        total_cost_cents: { type: integer, format: int64, nullable: true, minimum: 0 }
        notes: { type: string, nullable: true, maxLength: 500 }
        items:
          type: array
          minItems: 1
          maxItems: 20
          description: |
            Pelo menos um item. O mesmo item não pode aparecer duas vezes no mesmo registro.
          items: { $ref: "#/components/schemas/MaintenanceRecordItemRequest" }
    UpdateMaintenanceRecordRequest:
      type: object
      description: Não altera as linhas de item.
      properties:
        occurred_on: { type: string, format: date, nullable: true }
        mileage_km: { type: integer, format: int32, nullable: true, minimum: 0 }
        workshop_name: { type: string, nullable: true }
        total_cost_cents: { type: integer, format: int64, nullable: true, minimum: 0 }
        notes: { type: string, nullable: true }
    MaintenanceRecordItem:
      type: object
      required: [id, maintenance_item_id, item_slug, item_name]
      properties:
        id: { type: string, format: uuid }
        maintenance_item_id: { type: string, format: uuid }
        item_slug: { type: string }
        item_name: { type: string }
        description: { type: string, nullable: true }
        part_brand: { type: string, nullable: true }
        cost_cents: { type: integer, format: int64, nullable: true }
        warranty_months: { type: integer, format: int32, nullable: true }
        warranty_km: { type: integer, format: int32, nullable: true }
        warranty_until:
          type: string
          format: date
          nullable: true
          description: |
            Derivado da data do registro mais os meses de garantia. Nunca armazenado — o
            servidor calcula a cada leitura.
        warranty_until_km:
          type: integer
          format: int32
          nullable: true
          description: Derivado da quilometragem do registro mais os quilômetros de garantia.
    MaintenanceRecord:
      type: object
      required:
        [id, vehicle_id, occurred_on, mileage_km, kind, total_cost_cents, items, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        vehicle_id: { type: string, format: uuid }
        occurred_on: { type: string, format: date }
        mileage_km: { type: integer, format: int32 }
        kind: { $ref: "#/components/schemas/MaintenanceRecordKind" }
        workshop_name: { type: string, nullable: true }
        total_cost_cents: { type: integer, format: int64 }
        notes: { type: string, nullable: true }
        items:
          type: array
          items: { $ref: "#/components/schemas/MaintenanceRecordItem" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    MaintenanceRecordPage:
      type: object
      required: [data, next_cursor]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/MaintenanceRecord" }
        next_cursor: { type: string, nullable: true }

    # ------------------------------------------------------------ prazos
...